Modified the code to work with InBlockTransactions
Among other things implemented proper decryption of the default/mutant
transaction in core/crypto/validator_confidentiality.go
Where appropriate it was added directly an InBlockTransaction.
Some functions however where inherently dealing with the transaction type.
E.g. a function that executes the deploy of a transaction.
For those functions a converter from normal transaction to transactionSet
was created to interface with the rest of the code.
At this state wherever it is necessary to deal with a TransactionSet
only the first default is used.

Change-Id: I814cfe58421f67f81604b835eac53eee44207e2d
Signed-off-by: Ivan Puddu <puddui@student.ethz.ch>