Added the definition of the TxSetSpec
The txSetSpec is provided when the client calls the remote interface
of the peer. Before the client was calling different methods
with the same specification to indicate whether to deploy, invoke or
query. Now all this actions need to be specified in the tx set for
each transaction. Hence the definition of the type of action to
perform has been moved in the blockchainmessages.proto file.

Change-Id: I24582627e509b66e8b61779b673a9518b8280eb1
Signed-off-by: Ivan Puddu <puddui@student.ethz.ch>